fix: Change metadata getFunctions() and getProcedures() to ignore search_path visibility
Previously if no schema pattern was specified then a default filter was added to getFunctions()
and getProcedures() to restrict the returned set to only those that are visible to the user
via search_path.

This removes that filter so that calling either function with a null or empty schema pattern
will return all functions or procedures.