test: Use deterministic RNG for randomized test matrix
Change randomized test matrix to use a deterministic RNG. If the RNG_SEED environment
variable is defined, it will be used as the initial seed for the RNG. Alternatively,
if the GITHUB_PR_NUMBER environment variable is defined, it will be used to generate
a deterministic RNG seed to ensure that all test cycles for the same PR generate the
same test matrix of random environments.

If neither environment variable is defined then a random seed will be generated and
logged to the console. Re-running the action using that seed with the same matrix
generation codebase will generate the same matrix.

The actual RNG code is copied from seedrandom:
https://github.com/davidbau/seedrandom/commit/4460ad325a0a15273a211e509f03ae0beb99511a